Partially revert "external/openssl: add support for multilib builds"
The changes introduced by that commit (b95c4f932e1bb) don't work in the
NDK build (only the build system in the complete sources supports the arch
suffixes).

Replacing += with := is done to avoid compiler warning as with += every
architecture that is built adds another set of the same source files.